refactor generate module
