+++ utils/types.ts
@@ -1,0 +1,9 @@
+export type Data = {
+  success: boolean,
+  dataUrl: string | undefined
+}
+
+export type Coordinates = {
+  x: number,
+  y: number
+}
